chore: extract test utils for DI
