+++ suite-common/dependency-injection/package.json
@@ -9,5 +9,8 @@
         "depcheck": "yarn g:depcheck",
         "type-check": "yarn g:tsc --build",
         "test:unit": "yarn g:jest -c ../../jest.config.base.js"
+    },
+    "dependencies": {
+        "@trezor/type-utils": "workspace:*"
     }
 }
