+++ suite-common/dependency-injection/src/index.ts
@@ -1,1 +1,5 @@
 export { toGetter } from './toGetter';
+
+// Test
+export { mockNotExpected, mock } from './test/mock';
+export { createMockDeps } from './test/createMockDeps';
