+++ suite-common/dependency-injection/src/service.ts
@@ -1,0 +1,10 @@
+// Service can be a function
+export type ServiceFunction<TReturn, TParams extends any[]> = (...args: TParams) => TReturn;
+
+// or a static configuration
+export type DiConfiguration = string | number | boolean;
+
+// or object, where properties can be other services => recursive
+export type RecursiveDeps = {
+    [key: string]: ServiceFunction<any, any> | RecursiveDeps | DiConfiguration;
+};
