+++ suite-common/dependency-injection/src/test/mock.ts
@@ -1,0 +1,7 @@
+export const mock = <T extends (...args: any[]) => any>(fn: T) =>
+    jest.fn<ReturnType<T>, Parameters<T>>().mockImplementation(fn);
+
+export const mockNotExpected = <T extends (...args: any[]) => any>(key: string) =>
+    jest.fn<ReturnType<T>, Parameters<T>>().mockImplementation(() => {
+        throw new Error(`Method ${key} was not expected to be called`);
+    });
