+++ suite-common/suite-sync/src/data/labeling/tests/createUpdateOutputLabel.test.ts
@@ -1,3 +1,4 @@
+import { createMockDeps, mock } from '@suite-common/dependency-injection';
 import type { OutputTable } from '@suite-common/suite-sync-storage';
 import type { NetworkSymbol } from '@suite-common/wallet-config';
 import { asAccountDescriptor } from '@suite-common/wallet-types';
@@ -5,7 +6,6 @@ import type { StaticSessionId } from '@trezor/connect';
 import { err, ok } from '@trezor/type-utils';
 
 import { createSuiteSyncStorageMock } from '../../../../tests/createSuiteSyncStorageMock.mock';
-import { createMockDeps, mock } from '../../../../tests/utils';
 import { SuiteSyncUnavailableOnDeviceError } from '../../../createRefreshSuiteSyncKeys';
 import type { UpdateOutputLabelDeps } from '../createUpdateOutputLabel';
 import { createUpdateOutputLabel } from '../createUpdateOutputLabel';
