+++ suite-common/suite-sync/src/data/labeling/tests/createUpdateWalletLabel.test.ts
@@ -1,10 +1,10 @@
+import { createMockDeps, mock } from '@suite-common/dependency-injection';
 import type { WalletTable } from '@suite-common/suite-sync-storage';
 import { asWalletDescriptor } from '@suite-common/wallet-types';
 import { StaticSessionId } from '@trezor/connect';
 import { err, ok } from '@trezor/type-utils';
 
 import { createSuiteSyncStorageMock } from '../../../../tests/createSuiteSyncStorageMock.mock';
-import { createMockDeps, mock } from '../../../../tests/utils';
 import { SuiteSyncUnavailableOnDeviceError } from '../../../createRefreshSuiteSyncKeys';
 import { UpdateWalletLabelDeps, createUpdateWalletLabel } from '../createUpdateWalletLabel';
