+++ suite-common/suite-sync/tests/createSuiteSyncStorageMock.mock.ts
@@ -1,3 +1,4 @@
+import { createMockDeps, mockNotExpected } from '@suite-common/dependency-injection';
 import {
     AccountTable,
     AddressTable,
@@ -7,8 +8,6 @@ import {
     SuiteSyncTable,
     WalletTable,
 } from '@suite-common/suite-sync-storage';
-
-import { createMockDeps, mockNotExpected } from './utils';
 
 const createSuiteSyncTableMock = <T extends SuiteSyncTable<any>>(methods?: Partial<T>) =>
     createMockDeps<SuiteSyncTable<InferSuiteSyncTableEntity<T>>>({
